Trabajo-Final-UNLP-IoT/dockers/python/requirements.txt
paho-mqtt
opencv-python
numpy
PyYAML